+++ packages/dso-toolkit/src/components/panel/panel.models.ts
@@ -1,5 +1,6 @@
 export interface Panel<TemplateFnReturnType> {
   children: TemplateFnReturnType;
+  emphasized?: boolean;
   heading: TemplateFnReturnType;
   dsoCloseClick?: (e: CustomEvent) => void;
 }